+++ providers/capa/standard/capa_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/giantswarm/clustertest/pkg/wait"
 
 	"github.com/giantswarm/cluster-test-suites/common"
+	"github.com/giantswarm/cluster-test-suites/internal/state"
 )
 
 var _ = Describe("Common tests", func() {
@@ -19,10 +20,10 @@ var _ = Describe("Common tests", func() {
 
 	It("has all the control-plane nodes running", func() {
 		values := &application.ClusterValues{}
-		err := framework.MC().GetHelmValues(cluster.Name, cluster.Namespace, values)
+		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().Namespace, values)
 		Expect(err).NotTo(HaveOccurred())
 
-		wcClient, err := framework.WC(cluster.Name)
+		wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
 		Expect(err).NotTo(HaveOccurred())
 
 		Eventually(wait.Consistent(common.CheckControlPlaneNodesReady(wcClient, values.ControlPlane), 12, 5*time.Second)).
@@ -33,10 +34,10 @@ var _ = Describe("Common tests", func() {
 
 	It("has all the worker nodes running", func() {
 		values := &application.ClusterValues{}
-		err := framework.MC().GetHelmValues(cluster.Name, cluster.Namespace, values)
+		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().Namespace, values)
 		Expect(err).NotTo(HaveOccurred())
 
-		wcClient, err := framework.WC(cluster.Name)
+		wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
 		Expect(err).NotTo(HaveOccurred())
 
 		Eventually(wait.Consistent(common.CheckWorkerNodesReady(wcClient, values), 12, 5*time.Second)).