+++ providers/capv/upgrade/capv_test.go
@@ -1,0 +1,17 @@
+package upgrade
+
+import (
+	. "github.com/onsi/ginkgo/v2"
+
+	"github.com/giantswarm/cluster-test-suites/common"
+	"github.com/giantswarm/cluster-test-suites/internal/upgrade"
+)
+
+var _ = Describe("Basic upgrade test", Ordered, func() {
+	upgrade.Run()
+
+	// Finally run the common tests after upgrade is completed
+	common.Run(&common.TestConfig{
+		BastionSupported: false,
+	})
+})