+++ providers/capz/standard/capz_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/giantswarm/clustertest/pkg/wait"
 
 	"github.com/giantswarm/cluster-test-suites/common"
+	"github.com/giantswarm/cluster-test-suites/internal/state"
 )
 
 type ClusterValues struct {
@@ -27,10 +28,10 @@ var _ = Describe("Common tests", func() {
 
 	It("has all the control-plane nodes running", func() {
 		values := &ClusterValues{}
-		err := framework.MC().GetHelmValues(cluster.Name, cluster.Namespace, values)
+		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().Namespace, values)
 		Expect(err).NotTo(HaveOccurred())
 
-		wcClient, err := framework.WC(cluster.Name)
+		wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
 		Expect(err).NotTo(HaveOccurred())
 
 		Eventually(wait.Consistent(common.CheckControlPlaneNodesReady(wcClient, values.ControlPlane), 12, 5*time.Second)).
@@ -41,10 +42,10 @@ var _ = Describe("Common tests", func() {
 
 	It("has all the worker nodes running", func() {
 		values := &ClusterValues{}
-		err := framework.MC().GetHelmValues(cluster.Name, cluster.Namespace, values)
+		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().Namespace, values)
 		Expect(err).NotTo(HaveOccurred())
 
-		wcClient, err := framework.WC(cluster.Name)
+		wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
 		Expect(err).NotTo(HaveOccurred())
 
 		Eventually(wait.Consistent(CheckWorkerNodesReady(wcClient, values), 12, 5*time.Second)).