+++ CHANGELOG.md
@@ -7,14 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2023-08-21
+
+### Added
+
+- Add a PVC test.
+
+### Fixed
+
+- Bastion tests are disabled for CAPV.
+
+### Changed
+
+- Refactored state management to use singleton
+
+## [1.8.0] - 2023-08-14
+
 ### Added
 
 - Add common test to make sure all default apps are marked as `deployed`.
+- Upgrade cluster tests
 
 ### Changed
 
 - Bumped clustertest to `v0.1.1` to fix `GITHUB_TOKEN` issue.
 - Bumped clustertest to `v0.2.0` to use custom app values.
+
+### Added
+
+- Upgrade cluster tests
 
 ## [1.7.0] - 2023-07-20
 
@@ -97,7 +118,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example common tests
 - Dockerfile for running tests in CI
 
-[Unreleased]: https://github.com/giantswarm/cluster-test-suites/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-test-suites/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/giantswarm/cluster-test-suites/compare/v1.8.0...v1.9.0
+[1.8.0]: https://github.com/giantswarm/cluster-test-suites/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/giantswarm/cluster-test-suites/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/giantswarm/cluster-test-suites/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/giantswarm/cluster-test-suites/compare/v1.4.0...v1.5.0